parent-communication-bot/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.23.2
pydantic==2.9.0
python-dotenv==1.0.0
apscheduler==3.10.1
twilio==9.0.4
pyyaml==6.0.1
python-multipart==0.0.7
